refactor: transfer axiosInstance from App to its own file
